+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # CHANGELOG
 
+
+
+## v25.xx
+
+### Features
+- New output push notifications. (#117 v25.50-alpha)
+- Push notification supports tenant configuration. (#121 v26.02-alpha)
+- Delegated Email MS365 support. (#119 v26.04-alpha)
+
 ## v25.47
 
 ### Features